v11.1.0-rc2 release

# -----BEGIN PGP SIGNATURE-----
#
# iQEzBAABCgAdFiEEhpWov9P5fNqsNXdanKSrs4Grc8gFAmpp68cACgkQnKSrs4Gr
# c8hpwwgAqT1NmI+/Upq3EN341HSYW4qx3fTcXrk7UgGDZFLv7Aldt6/DwY2FEzvQ
# ZOEoehclwmCQZL0itvDQD5eZQDrQcz1ar9HeFm3CzDwDh++YAUsJS0+TMaHQJOwq
# jk2H/tyVH51y9xFAzNzoC1z5qb8lN5+LlJJABbkaPe/KLoJvwVqyxLNAZQg/Aob8
# iSVdR+PDiSjYFtA+A1BE1RGr/mqIIKctM8gNAQM0gr3ZFmXFCfzgnMFRBJLcybGC
# HVAproGcCZ3yCT30/jHb9Job61dI0L+Sdb6JLIofWcMAE23VRPblxhlSS3f+SKwV
# Davi5Mf/Mbh9siA/PlHNjOkLiyIn3g==
# =ZgXp
# -----END PGP SIGNATURE-----
# gpg: Signature made Wed Jul 29 15:02:15 2026 +03
# gpg:                using RSA key 8695A8BFD3F97CDAAC35775A9CA4ABB381AB73C8
# gpg: Can't check signature: No public key

# Conflicts:
#	.gitlab-ci.d/base.yml
#	.gitlab-ci.d/macos.yml
#	VERSION
#	disas/riscv.c
#	hw/9pfs/9p.c
#	hw/arm/meson.build
#	hw/i386/amd_iommu.c
#	hw/i3c/dw-i3c.c
#	hw/misc/aspeed_hace.c
#	hw/nvme/ctrl.c
#	hw/vfio/container.c
#	linux-user/strace.c
#	linux-user/strace.list
#	linux-user/syscall.c
#	migration/migration.c
#	migration/vmstate.c
#	roms/SLOF
#	roms/openbios
#	roms/opensbi
#	target/arm/hvf/hvf.c
#	target/arm/tcg/translate-sme.c
#	target/arm/tcg/translate-sve.c
#	target/i386/tcg/decode-new.c.inc
#	target/i386/tcg/emit.c.inc
#	target/riscv/machine.c
#	target/riscv/tcg/cpu_helper.c
#	target/riscv/tcg/csr.c
#	target/riscv/tcg/tcg-cpu.c
#	target/riscv/tcg/vector_helper.c
#	tests/qtest/virtio-9p-test.c
#	ui/console-vc.c

+++ hw/sparc64/niagara.c
@@ -25,7 +25,7 @@
 #include "qemu/osdep.h"
 #include "block/block_int-common.h"
 #include "qemu/units.h"
-#include "cpu.h"
+#include "target/sparc/cpu.h"
 #include "hw/core/boards.h"
 #include "hw/char/serial-mm.h"
 #include "hw/misc/unimp.h"
@@ -137,7 +137,7 @@ static void niagara_init(MachineState *machine)
        outside of the partition RAM */
     if (dinfo) {
         BlockBackend *blk = blk_by_legacy_dinfo(dinfo);
-        int size = blk_getlength(blk);
+        int64_t size = blk_getlength(blk);
         if (size > 0) {
             memory_region_init_ram(&s->vdisk_ram, NULL, "sun4v_vdisk.ram", size,
                                    &error_fatal);
